Updated to use baseapi and modules

+++ app/src/main/java/com/churchclerk/churchapi/ChurchApi.java
@@ -4,6 +4,7 @@
 package com.churchclerk.churchapi;
 
 import com.churchclerk.baseapi.BaseApi;
+import com.churchclerk.baseapi.model.ApiCaller;
 import com.churchclerk.churchapi.model.Church;
 import com.churchclerk.churchapi.service.ChurchService;
 import org.slf4j.Logger;
@@ -12,9 +13,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Component;
+import org.springframework.web.client.HttpClientErrorException;
 
-import javax.ws.rs.Path;
+import javax.ws.rs.*;
 import java.util.Date;
+import java.util.UUID;
 
 /**
  *
@@ -25,6 +28,10 @@ public class ChurchApi extends BaseApi<Church> {
 
     private static Logger logger = LoggerFactory.getLogger(ChurchApi.class);
 
+    @QueryParam("name")
+    protected String nameLike;
+
+
     @Autowired
     private ChurchService service;
 
@@ -33,9 +40,14 @@ public class ChurchApi extends BaseApi<Church> {
      */
     public ChurchApi() {
         super(logger, Church.class);
-        setReadRoles(Role.ADMIN, Role.CLERK, Role.OFFICIAL, Role.MEMBER, Role.NONMEMBER);
-        setUpdateRoles(Role.ADMIN);
-        setDeleteRoles(Role.ADMIN);
+        setReadRoles(ApiCaller.Role.ADMIN, ApiCaller.Role.CLERK, ApiCaller.Role.OFFICIAL, ApiCaller.Role.MEMBER, ApiCaller.Role.NONMEMBER);
+        setUpdateRoles(ApiCaller.Role.ADMIN);
+    }
+
+    @Override
+    protected Page<? extends Church> doGet(Pageable pageable) {
+
+        return service.getResources(pageable, createCriteria());
     }
 
     /**
@@ -47,45 +59,97 @@ public class ChurchApi extends BaseApi<Church> {
 
         addBaseCriteria(criteria);
 
+        criteria.setName(nameLike);
+
+        if ((id == null) && (hasSuperRole() == false)) {
+            criteria.setId("NOTALLOWED");
+        }
+
+        if (readAllowed(id, this::hasSuperRole)) {
+            if (id != null) {
+                criteria.setId(id);
+            }
+        }
+        else {
+            // force return of empty array
+            criteria.setId("NOTALLOWED");
+        }
+
         return criteria;
     }
 
     @Override
-    protected Page<? extends Church> doGet(Pageable pageable) {
+    protected Church doGet() {
+        if ((id == null) || (id.trim().isEmpty())) {
+            throw new BadRequestException("Church id cannot be empty");
+        }
 
-        return service.getResources(pageable, createCriteria());
-    }
+        Church resource = service.getResource(id);
 
-    @Override
-    protected Church doGet(String id) {
+        if ((resource == null) || (readAllowed(resource.getId()) == false)) {
+            throw new NotFoundException();
+        }
 
-            Church resource = service.getResource(id);
-
-            return resource;
-    }
-
-    @Override
-    protected Church doCreate(Church resource) {
-
-        resource.setCreatedBy(requesterId);
-        resource.setCreatedDate(new Date());
-        resource.setUpdatedBy(requesterId);
-        resource.setUpdatedDate(new Date());
         return resource;
     }
 
     @Override
-    protected Church doUpdate(Church resource) {
-        Church found = service.getResource(id);
+    protected Church doCreate(Church resource) {
+        if (resource.getId() != null) {
+            throw new BadRequestException("Church id should not be present");
+        }
 
-        resource.setUpdatedBy(requesterId);
+        if (resource.getName() == null) {
+            throw new BadRequestException("Church's name cannot be null");
+        }
+
+        if (createAllowed(resource.getId(), this::hasSuperRole) == false) {
+            throw new ForbiddenException();
+        }
+
+        resource.setId(UUID.randomUUID().toString());
+        resource.setCreatedBy(apiCaller.getUserid());
+        resource.setCreatedDate(new Date());
+        resource.setUpdatedBy(apiCaller.getUserid());
+        resource.setUpdatedDate(new Date());
+
+        return service.createResource(resource);
+    }
+
+    @Override
+    protected Church doUpdate(Church resource) {
+        if ((id == null) || (id.isEmpty()) || (resource.getId() == null) || (resource.getId().isEmpty())) {
+            throw new BadRequestException("Church id cannot be empty");
+        }
+
+        if (resource.getId().equals(id) == false) {
+            throw new BadRequestException("Church id does not match");
+        }
+
+        if ((resource.getName() == null) || (resource.getName().trim().isEmpty())) {
+            throw new BadRequestException("Church name cannot be empty");
+        }
+
+        if (updateAllowed(id) == false) {
+            throw new ForbiddenException();
+        }
+
+        resource.setUpdatedBy(apiCaller.getUserid());
         resource.setUpdatedDate(new Date());
 
         return service.updateResource(resource);
     }
 
     @Override
-    protected Church doDelete(String id) {
+    protected Church doDelete() {
+        if ((id == null) || id.isEmpty()) {
+            throw new BadRequestException("Church id cannot be empty");
+        }
+
+        if (deleteAllowed(id) == false) {
+            throw new ForbiddenException();
+        }
+
         return service.deleteResource(id);
     }
 }
